+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed #15 consensus on the threshold of 1/3 sleeping validators. (#388)
 - Replaced config param `timeout_events_capacity` with `internal_events_capacity`. (#388)
 - The `Transaction` trait now inherit `ExonumJson`. (#402)
+- The list of peer connections is now restored to the last state after the process is restarted. (#378)
 - `Transaction` `execute` method now returns  `TransactionStatus`. (#385)
 
 ### Removed